+++ CHANGELOG.md
@@ -1,12 +1,14 @@
 [1.10.0] - ????-??-??
 
+Changed:
+* Ctrl-c now causes a graceful shutdown and commits the database journal before closing it.
+
 Fixed:
 * A small fix to the scroll wheel behavior on the puzzle board to prevent the page from
   scrolling if the mousewheel is over the board.
 * Fix countdown timer to next review not starting sometimes.
-
-Changed:
-* Ctrl-c now causes a graceful shutdown and commits the database journal before closing it.
+* The 'next review due' time on the stats screen no longer says 'never' when there are no reviews
+  due for the rest of the day.
 
 [1.9.0] - 2023-10-26